% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{authorClass-class}
\alias{authorClass}
\alias{authorClass-class}
\title{S4 class for representing author's information}
\description{
S4 class represents information of an author of a \code{rrisk} model
}
\section{Slots}{

\describe{
\item{\code{name}}{name}

\item{\code{institution}}{institution}

\item{\code{email}}{email address}
}}
\examples{
\donttest{new("authorClass")

new("authorClass",name="Max Mustermann",institution="Example institution",
email="example@email.com")}
}

