% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{basicsClass-class}
\alias{basicsClass}
\alias{basicsClass-class}
\title{S4 class for representing basic descriptions}
\description{
S4 class for representing basic descriptions of a \code{rrisk} model
}
\examples{
\donttest{new("basicsClass")

new("basicsClass",
name="Background",
explanation=gsub(x="Escherichia (E.) coli O157:H7 represents a microbial 
hazard in food stuffs such as ground beef. It is known that good kitchen
hygiene and proper cooking reduces the risk of food-borne outbreaks
caused by this bacterium. A model for assessing the risk of illness
in children less than...","\\n",replacement=""))}
}

