% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{bsrv.fittodata}
\alias{bsrv.fittodata}
\title{Non-executable auxiliary function}
\usage{
bsrv.fittodata(data2fitname,item,rriskModel,method)
}
\arguments{
\item{data2fitname}{...}

\item{item}{...}

\item{rriskModel}{...}

\item{method}{character string giving the method for bootstrapping, \code{method=c("param","nonparam")}}
}
\keyword{items}

