% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeFunctions.R
\name{change.settings}
\alias{change.settings}
\title{Function for edditing model slot 'settings'}
\usage{
change.settings(settings)
}
\arguments{
\item{settings}{(model settings) is an object of class \code{modelSettingsClass}}
}
\value{
This function returns edited model slot 'settings'.
}
\description{
A change function for \code{\linkS4class{modelSettingsClass}}.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\note{
This function requires some functions from the package \code{tcltk}.
}
\examples{
\donttest{model<-init.Model1()
model@settings<-change.settings(model@settings)
model@settings}
}
\keyword{edit}

