% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexFunctions.R
\name{changeUpdate}
\alias{changeUpdate}
\title{Non-executable auxiliary function}
\usage{
changeUpdate(rriskModel,menuLevel=1)
}
\arguments{
\item{rriskModel}{...}

\item{menuLevel}{...}
}
\keyword{...}
\keyword{report}

