% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexFunctions.R
\name{changeUpdateAuthor}
\alias{changeUpdateAuthor}
\title{Non-executable auxiliary function}
\usage{
changeUpdateAuthor(rriskModel,menuLevel=1)
}
\arguments{
\item{rriskModel}{...}

\item{menuLevel}{...}
}
\keyword{report}

