% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{database}
\alias{NorthSouth}
\alias{database}
\alias{datasets}
\alias{sdata}
\alias{zipdata}
\title{Data sets}
\examples{
\donttest{
data(sdata)
sdata

data(zipdata)
zipdata

data(NorthSouth)
NorthSouth
}
}
\keyword{dataset}

