\name{define.fdoc}
\alias{define.fdoc}
\title{Non-executable auxiliary function}
\usage{
  define.fdoc(item)
}
\arguments{
  \item{item}{...}
}
\description{
  Function that loads external models
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\keyword{items}

