\name{define.mxrv.stramcrv}
\alias{define.mxrv.stramcrv}
\title{Non-executable auxiliary function}
\usage{
  define.mxrv.stramcrv(item,rriskModel,mcrv.item.name)
}
\arguments{
  \item{item}{...}

  \item{rriskModel}{...}

  \item{mcrv.item.name}{...}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

