\name{define.numv.fnrv}
\alias{define.numv.fnrv}
\title{Non-executable auxiliary function}
\usage{
  define.numv.fnrv(item,rriskModel,menuLevel=1,type="numv")
}
\arguments{
  \item{item}{...}

  \item{rriskModel}{an instance of the class
  \code{modelClass}}

  \item{menuLevel}{...}

  \item{type}{character string defining item type,
  \code{type=c("numv","fnrv")}}
}
\description{
  Function that evaluates and saves the input numeric value
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\keyword{items}

