% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{glossaryClass-class}
\alias{glossaryClass}
\alias{glossaryClass-class}
\title{S4 class for representing model glossary entry}
\description{
S4 class for representing a glossary of a \code{rrisk} model
}
\examples{
\donttest{new("glossaryClass")}
}

