% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{graphClass-class}
\alias{graphClass}
\alias{graphClass-class}
\title{S4 class for representing model graph object}
\description{
S4 class for representing a graph used in a \code{rrisk} model
}
\examples{
 \donttest{new("graphClass")}
}

