% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFunctions.R
\name{init.Model1}
\alias{init.Model1}
\title{Function to initialize an object as the instance of 'modelClass'}
\usage{
init.Model1(demomode=FALSE)
}
\arguments{
\item{demomode}{is a boolean value indicating whether the model should be run under the demo mode}
}
\value{
demoModel1 auxiliary boolean, \code{demomode=c(TRUE,FALSE)}
}
\description{
This initializes an object as the instance of \code{\linkS4class{modelClass}}.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskModel<-init.Model1()}
}
\keyword{init}

