\name{itemsEvaluation}
\alias{itemsEvaluation}
\title{Non-executable auxiliary function}
\usage{
  itemsEvaluation(item,rriskModel,run=FALSE,catEval=FALSE)
}
\arguments{
  \item{item}{...}

  \item{rriskModel}{...}

  \item{run}{a single logical value}

  \item{catEval}{a single logical value}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

