\name{mcrv.fittodata}
\alias{mcrv.fittodata}
\title{Non-executable auxiliary function}
\usage{
  mcrv.fittodata(data2fit,data2fitname,item,rriskModel)
}
\arguments{
  \item{data2fit}{...}

  \item{data2fitname}{...}

  \item{item}{...}

  \item{rriskModel}{...}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

