% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{menu.bdjp.ZIPGUI}
\alias{menu.bdjp.ZIPGUI}
\title{Non-executable auxiliary function}
\usage{
menu.bdjp.ZIPGUI(item,itemdef,rriskModel,menuLevel=1)
}
\arguments{
\item{item}{...}

\item{itemdef}{...}

\item{rriskModel}{...}

\item{menuLevel}{...}
}
\keyword{items}

