% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuFunctions.R
\name{menu.copySsystem}
\alias{menu.copySsystem}
\title{Function listing all available scoring systems and offering them to be duplicated}
\usage{
menu.copySsystem(rriskSessionSsystems,menuLevel=1)
}
\arguments{
\item{rriskSessionSsystems}{is the slot scoringsystem from an instance of \code{rriskClass}}

\item{menuLevel}{is a numeric value that indicates the load of the current menu}
}
\description{
This function lists all available scoring systems and offers them to be duplicated.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskSession<-init.rriskSession(useDemoModels = "all")
menu.copySsystem(rriskSession@scoringsystems,menuLevel=1)}
}
\keyword{menu}

