\name{menu.define.bsrv}
\alias{menu.define.bsrv}
\title{Non-executable auxiliary function}
\usage{
  menu.define.bsrv(item,rriskModel,menuLevel=1)
}
\arguments{
  \item{item}{...}

  \item{rriskModel}{...}

  \item{menuLevel}{...}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

