\name{menu.define.data}
\alias{menu.define.data}
\title{Non-executable auxiliary function}
\usage{
  menu.define.data(item,menuLevel=1)
}
\arguments{
  \item{item}{...}

  \item{menuLevel}{numeric value indicating menu level}
}
\description{
  This function generates the user menu for maintaining
  scoring systems
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\keyword{items}

