% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuFunctions.R
\name{menu.handleModel}
\alias{menu.handleModel}
\title{Function generating user menu for maintaining models}
\usage{
menu.handleModel(rriskModel,rriskSession,menuLevel=1)
}
\arguments{
\item{rriskModel}{is an instance of the \code{modelClass}}

\item{rriskSession}{is an instance of the \code{rriskClass}}

\item{menuLevel}{is a numeric value that indicates the load of the current menu}
}
\description{
This function generates the user menu for maintaining models.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskSession<-init.rriskSession(useDemoModels = "all")
rriskModel.new<-menu.handleModel(rriskSession@models[[1]],rriskSession)}
}
\keyword{menu}

