% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{menu.mcrv.dfamily}
\alias{menu.mcrv.dfamily}
\title{Non-executable auxiliary function}
\usage{
menu.mcrv.dfamily(item,rriskModel,dfamily="pois",menuLevel=1)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{dfamily}{...}

\item{menuLevel}{...}
}
\keyword{items}

