\name{menu.mcrv.fittodata}
\alias{menu.mcrv.fittodata}
\title{Non-executable auxiliary function}
\usage{
  menu.mcrv.fittodata(item,rriskModel,menuLevel=1)
}
\arguments{
  \item{item}{...}

  \item{rriskModel}{...}

  \item{menuLevel}{...}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

