% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuFunctions.R
\name{menu.runModel}
\alias{menu.runModel}
\title{Function generating user menu for running simulations on the model}
\usage{
menu.runModel(rriskModel,menuLevel=1)
}
\arguments{
\item{rriskModel}{is an instance of the \code{modelClass}}

\item{menuLevel}{is a numeric value that indicates the load of the current menu}
}
\description{
This function generates the user menu for running simulations on the model.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskModel<-init.Model1()
menu.runModel(rriskModel)}
}
\keyword{menu}

