% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuFunctions.R
\name{menu.runSensitivityModel}
\alias{menu.runSensitivityModel}
\title{Function generating user menu for viewing graphics}
\usage{
menu.runSensitivityModel(rriskModel,menuLevel=1)
}
\arguments{
\item{rriskModel}{is an instance of the \code{modelClass}}

\item{menuLevel}{is a numeric value that indicates the load of the current menu}
}
\description{
This function generates the user menu for viewing graphics.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskModel<-init.Model1()   
menu.runSensitivityModel(rriskModel)}
}
\keyword{menu}

