% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuFunctions.R
\name{menu.saveModel}
\alias{menu.saveModel}
\title{Function generating user menu for saving models}
\usage{
menu.saveModel(rriskSessionModels,menuLevel=1)
}
\arguments{
\item{rriskSessionModels}{are the slot models from the \code{rriskClass}}

\item{menuLevel}{is a numeric value that indicates the load of the current menu}
}
\description{
This function generates the user menu for saving models.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskSession<-init.rriskSession(useDemoModels="all")
menu.saveModel(rriskSession@models,menuLevel=1)}
}
\keyword{menu}

