% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuFunctions.R
\name{menu.viewItems}
\alias{menu.viewItems}
\title{Non-executable auxiliary function}
\usage{
menu.viewItems(rriskModel,menuLevel=1)
}
\arguments{
\item{rriskModel}{...}

\item{menuLevel}{...}
}
\keyword{menu}

