% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelConclusionsClass-class}
\alias{modelConclusionsClass}
\alias{modelConclusionsClass-class}
\title{S4 class for representing all 'rrisk' model conclusions}
\description{
S4 class for representing all the conclusions of a \code{rrisk} model
}
\examples{
 \donttest{new("modelConclusionsClass")

new("modelConclusionsClass",
conclusions=list(gsub(x="This illustrative model predicts that the population
average probability of illness due to E.coli O157:H7 in child below 
age of three years attributable to consumption of one serving prepared 
from ground beef is around...","\\n",replacement=""),
gsub(x="The probabilistic modelling approach indicates that, taking into 
account variability and uncertainty, this probability could be much 
higher in rare cases...","\\n",replacement=""))) }
}

