% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelGraphsClass-class}
\alias{modelGraphsClass}
\alias{modelGraphsClass-class}
\title{S4 class for representing all 'rrisk' model graphs objects}
\description{
S4 class for representing all the graphs used in \code{rrisk} models
}
\examples{
 \donttest{new("modelGraphsClass")}
}

