% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelNameClass-class}
\alias{modelNameClass}
\alias{modelNameClass-class}
\title{S4 class for representing the names of 'rrisk' models}
\description{
A S4 class for representing the names of \code{rrisk} models
}
\examples{
\donttest{new("modelNameClass")}
}

