% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelSettingsClass-class}
\alias{modelSettingsClass}
\alias{modelSettingsClass-class}
\title{S4 class for representing 'rrisk' model settings}
\description{
S4 class for representing information of all the default settings of a \code{rrisk} model
}
\examples{
\donttest{show(new("modelSettingsClass"))}
}

