% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{mymenu}
\alias{mymenu}
\title{Function that produces menu element and catches the input of the user}
\usage{
mymenu(title="Hi",choices="y",codes=1:length(choices),part="NA",
   help="No further help available",levelTabulator)
}
\arguments{
\item{title}{the title}

\item{choices}{choice made my the user}

\item{codes}{vector from 1 to the length of choices}

\item{part}{part}

\item{help}{whether further help is available}

\item{levelTabulator}{numeric indicating the level of the menu}
}
\description{
Function mymenu that produces menu element and catches the input of the user
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{mymenu(title="Hi",choices="x",codes=1:length("x"),
part="NA",help="No further help available",levelTabulator=1)}
}
\keyword{misc}

