% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{myvalue.distrparams}
\alias{myvalue.distrparams}
\title{Non-executable auxiliary function}
\usage{
myvalue.distrparams(item,rriskModel,
       title="Tryme",menuLevel=1,
       minmax,asInteger=FALSE,batch=1)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{title}{...}

\item{menuLevel}{...}

\item{minmax}{...}

\item{asInteger}{boolean value giving whether entered value should be an integer, \code{asInteger=c(TRUE,FALSE)}}

\item{batch}{...}
}
\keyword{items}

