% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{myvalue.stratumNames}
\alias{myvalue.stratumNames}
\title{Non-executable auxiliary function}
\usage{
myvalue.stratumNames(item,rriskModel,menuLevel=1,stratumNames)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{menuLevel}{...}

\item{stratumNames}{...}
}
\keyword{items}

