\name{numericNULL-class}
\Rdversion{1.1}
\docType{class}
\alias{numericNULL-class}

\title{S4 virtual class}
\description{A S4 virtual class, that means a union of classes \code{numeric} and \code{NULL}.}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "numericNULL" in the signature.
}

\keyword{classes}

