% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{partClass-class}
\alias{partClass}
\alias{partClass-class}
\title{S4 class for representing model part item}
\description{
S4 class for representing descriptions of a part of a \code{rrisk} model
}
\examples{
\donttest{new("partClass")}
}

