\name{rdiscrete}
\alias{rdiscrete}
\title{rdiscrete}
\usage{
  rdiscrete(n,values,probs)
}
\arguments{
  \item{n}{n}

  \item{values}{values}

  \item{probs}{probs}
}
\description{
  Function rdiscrete
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\keyword{manip}

