\name{referencesLoad}
\alias{referencesLoad}
\title{Function loads external reference information}
\usage{
  referencesLoad()
}
\description{
  Function loads external reference information
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\examples{
\donttest{referencesLoad()}
}
\keyword{model}

