\docType{package}
\name{rrisk-package}
\alias{rrisk}
\alias{rrisk-package}
\title{'rrisk' project.}
\description{
  'rrisk' project.
}
\details{
  \tabular{ll}{ Package: \tab rrisk\cr Type: \tab
  Package\cr Version: \tab 2.11\cr Date: \tab 2012-10-26\cr
  Depends: \tab R(>= 2.11.0), methods, tcltk, png, rgdal,
  grDevices, network, sna, tools, tree, mgcv, rriskBayes,
  rriskDistributions, xtable, fitdistrplus\cr License: \tab
  GPL(>=3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://www.bfr.bund.de/cd/52158\cr }

  This R packages provides the whole functionality of the
  rrisk project developed by Federal Institute for Risk
  Assessment. This tool supports the risk assessor in
  definition, implementation and documantation of
  quantitative risk models.
}
\author{
  Matthias Greiner \email{matthias.greiner@bfr.bund.de},
  Natalia Belgorodski \email{belgorodski@stat-up.com},
  Kristin Tolksdorf, Christine Mueller-Graf, Katharina
  Schueller \email{schueller@stat-up.com}, Yinchong Yang
  \email{yang@stat-up.com}, Markus Junker
}
\keyword{package}

