\name{rrisksample}
\alias{rrisksample}
\title{Non-executable auxiliary function}
\usage{
  rrisksample(data,weights)
}
\arguments{
  \item{data}{data frame}

  \item{weights}{vector with weights for sampling from data
  frame columns}
}
\description{
  Sampling from dara frame column by predefined weights
}
\keyword{manip}

