% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{rsrv.resample}
\alias{rsrv.resample}
\title{Non-executable auxiliary function}
\usage{
rsrv.resample(data2fitname,item,rriskModel)
}
\arguments{
\item{data2fitname}{...}

\item{item}{...}

\item{rriskModel}{...}
}
\keyword{items}

