\name{rudiscrete}
\alias{rudiscrete}
\title{rudiscrete}
\usage{
  rudiscrete(n=100,min,max)
}
\arguments{
  \item{n}{n}

  \item{min}{min}

  \item{max}{max}
}
\description{
  Function rudiscrete
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\keyword{manip}

