% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFunctions.R
\name{run2dsimulation}
\alias{run2dsimulation}
\title{Non-executable auxiliary function}
\usage{
run2dsimulation(rriskModel,menuLevel=1)
}
\arguments{
\item{rriskModel}{...}

\item{menuLevel}{...}
}
\keyword{run}

