% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{scoreClass-class}
\alias{scoreClass}
\alias{scoreClass-class}
\title{S4 class for representing model score items}
\description{
S4 class for representing information of a scoring system of a \code{rrisk} model
}
\examples{
 \donttest{new("scoreClass")}
}

