% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumupMethods.R
\docType{methods}
\name{sumUp}
\alias{sumUp}
\alias{sumUp,modelClass-method}
\alias{sumUp,modelScoringClass-method}
\alias{sumUp,rriskClass-method}
\alias{sumUp-methods}
\title{Generic function and methods for object summing up}
\usage{
sumUp(object,...)

\S4method{sumUp}{modelClass}(object, showFullCommand = FALSE)

\S4method{sumUp}{modelScoringClass}(object)

\S4method{sumUp}{rriskClass}(object)
}
\arguments{
\item{object}{An object}

\item{object}{an \code{\linkS4class{modelClass}} object}

\item{object}{an \code{\linkS4class{modelScoringClass}} object}

\item{object}{an \code{\linkS4class{rriskClass}} object}
}
\description{
Generic function for summing up objects

\code{sumUp} method for \code{\linkS4class{modelClass}}

\code{sumUp} method for \code{\linkS4class{modelScoringClass}}

\code{sumUp} method for \code{\linkS4class{rriskClass}}
}
\examples{
\donttest{sumUp(new("modelClass"))
sumUp(init.Model1())}
\donttest{sumUp(new("modelScoringClass"))
sumUp(init.scoresystem1())}
\donttest{sumUp(new("rriskClass"))
sumUp(init.rriskSession())}
}

