% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexFunctions.R
\name{tolatex.basics}
\alias{tolatex.basics}
\title{LaTeX code producing the general description of the model}
\usage{
tolatex.basics(rriskModel, file.name)
}
\arguments{
\item{rriskModel}{is an instance of the \code{rriskModel}}

\item{file.name}{is a character value that indicates the Tex file where codes to be written into}
}
\description{
This creates the LaTeX scripts producing the first part in the documentation - 
"General Description of the model".
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskModel<-init.Model1()
tolatex.basics(rriskModel=rriskModel, file.name="report.tex") }
}
\keyword{report}

