% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexFunctions.R
\name{tolatex.frontmatter}
\alias{tolatex.frontmatter}
\title{LaTeX code for setting the frontmatter}
\usage{
tolatex.frontmatter(rriskModel, disclaimer, file.name)
}
\arguments{
\item{rriskModel}{is an instance of the \code{rriskModel}}

\item{disclaimer}{is the slot disclaimer of an instance of \code{rriskClass}}

\item{file.name}{is a character value that indicates the Tex file where codes to be written into}
}
\description{
This creates the LaTeX scripts producing the entire frontmatter of the documentation 
of 5 pages, including the coverpage, inside front cover, contents, acknowledgements 
as well as a brief introduction of the document.
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskModel<-init.Model1()
rriskSession<- init.rriskSession(useDemoModels = "all")
tolatex.frontmatter(rriskModel=rriskModel, disclaimer=rriskSession@disclaimer, 
file.name="report.tex") }
}
\keyword{report}

