% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexFunctions.R
\name{tolatex.modelResults}
\alias{tolatex.modelResults}
\title{LaTeX code for producing texts displaying the model results}
\usage{
tolatex.modelResults(rriskModel, file.name)
}
\arguments{
\item{rriskModel}{is an instance of the \code{rriskModel}}

\item{file.name}{is a character value that indicates the Tex file where codes to be written into}
}
\description{
This creates the LaTeX scripts producing the fifth part in the documentation - 
"model results ".
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskModel<-init.Model1()
tolatex.modelResults(rriskModel=rriskModel, file.name="report.tex") }
}
\keyword{report}

