% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexFunctions.R
\name{tolatex.scoringTable}
\alias{tolatex.scoringTable}
\title{producing LaTeX codes tabling the scoring system of the model with the
default criteria and qualitive scores
required by the model}
\usage{
tolatex.scoringTable(rriskModel,file.name)
}
\arguments{
\item{rriskModel}{is an instance of the \code{rriskModel}}

\item{file.name}{is a character value that indicates the Tex file where codes to be written into}
}
\description{
This creates the LaTeX scripts for the first table in the appendix - 
"Scoring table".
}
\details{
This function is not intended to be called directly but is internally called
during \code{rrisk} session.
}
\examples{
\donttest{rriskModel<-init.Model1()
tolatex.scoringTable(rriskModel=rriskModel,file.name="report.tex")}
}
\keyword{report}

