% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{uncertClass-class}
\alias{uncertClass}
\alias{uncertClass-class}
\title{S4 class for representing model uncertainty item}
\description{
S4 class for representing descriptions of an uncertainty of a \code{rrisk} model
}
\section{Slots}{

\describe{
\item{\code{namemain}}{main category of uncertainty}

\item{\code{namesub}}{sub category of uncertainty}
}}
\examples{
\donttest{new("uncertClass")}
}

