% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeoutMethods.R
\docType{methods}
\name{writeOut}
\alias{writeOut}
\alias{writeOut,modelClass-method}
\alias{writeOut,modelReferencesClass-method}
\alias{writeOut,modelScoringClass-method}
\alias{writeOut-methods}
\title{Generic function and methods for writing specific model informations into a file}
\usage{
writeOut(object, ...)

\S4method{writeOut}{modelReferencesClass}(object)

\S4method{writeOut}{modelClass}(object)

\S4method{writeOut}{modelScoringClass}(object)
}
\arguments{
\item{object}{An object}

\item{model.name}{name of the model}

\item{object}{a \code{modelReferencesClass} object}

\item{object}{a \code{modelClass} object}

\item{object}{a \code{modelScoringClass} object}
}
\description{
Generic function for writing model informations list into a file

writeOut method for \code{\linkS4class{modelReferencesClass}}

writeOut method for \code{\linkS4class{modelClass}}

writeOut method for \code{\linkS4class{modelScoringClass}}
}
\examples{
\donttest{writeOut(new("modelReferencesClass"))}
\donttest{
writeOut(new("modelClass"))
writeOut(init.Model1())}
\donttest{writeOut(new("modelScoringClass"))
writeOut(init.scoresystem1())}
}

