% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{activateFDOCs}
\alias{activateFDOCs}
\title{Non-executable auxiliary function}
\usage{
activateFDOCs(items)
}
\arguments{
\item{items}{...}
}
\description{
This function activates all functions documented in 'fdoc' items, 
i.e. this function add all fdoc function to the workspace.
}
\keyword{items}

