\name{bdjp.saveResults}
\alias{bdjp.saveResults}
\title{Non-executable auxiliary function}
\usage{
  bdjp.saveResults(bayesResults,item,itemdef="",rriskModel,menuLevel=1)
}
\arguments{
  \item{bayesResults}{...}

  \item{item}{...}

  \item{itemdef}{...}

  \item{rriskModel}{...}

  \item{menuLevel}{...}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

