\name{bsrv.fittodata}
\alias{bsrv.fittodata}
\title{Non-executable auxiliary function}
\usage{
  bsrv.fittodata(data2fitname,item,rriskModel,method)
}
\arguments{
  \item{data2fitname}{...}

  \item{item}{...}

  \item{rriskModel}{...}

  \item{method}{character string giving the method for
  bootstrapping, \code{method=c("param","nonparam")}}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

