\name{dataLoad}
\alias{dataLoad}
\title{Non-executable auxiliary function}
\usage{
  dataLoad(item)
}
\arguments{
  \item{item}{...}
}
\description{
  Function that loads data sets
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\keyword{items}

