% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{define.mxrv.stramcrv}
\alias{define.mxrv.stramcrv}
\title{Non-executable auxiliary function}
\usage{
define.mxrv.stramcrv(item,rriskModel,mcrv.item.name)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{mcrv.item.name}{...}
}
\keyword{items}

