% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{get.dataForWith}
\alias{get.dataForWith}
\title{Non-executable auxiliary function}
\usage{
get.dataForWith(item,rriskModel,catResults=FALSE,fullc=TRUE,run=FALSE,run2d=FALSE,catEval=TRUE)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{catResults}{...}

\item{fullc}{...}

\item{run}{...}

\item{run2d}{...}

\item{catEval}{...}
}
\keyword{items}

