\name{graphLoad}
\alias{graphLoad}
\title{Function that loads external graphics}
\usage{
  graphLoad()
}
\description{
  Function that loads external graphics
}
\details{
  This function is not intended to be called directly but
  is internally called during \code{rrisk} session.
}
\examples{
\donttest{graphLoad()}
}
\keyword{model}

