% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{itemClass-class}
\alias{itemClass}
\alias{itemClass-class}
\title{S4 class for representing model items}
\description{
S4 class for representing an Item of \code{rrisk} models
}
\examples{
\donttest{new("itemClass")}
}

