% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{itemsEvaluation}
\alias{itemsEvaluation}
\title{Non-executable auxiliary function}
\usage{
itemsEvaluation(item,rriskModel,run=FALSE,catEval=FALSE)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{run}{a single logical value}

\item{catEval}{a single logical value}
}
\keyword{items}

