% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{menu.mcrv.cfamily}
\alias{menu.mcrv.cfamily}
\title{Non-executable auxiliary function}
\usage{
menu.mcrv.cfamily(item,rriskModel,dfamily="beta",menuLevel=1)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{dfamily}{...}

\item{menuLevel}{...}
}
\keyword{items}

