% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineItemtype.R
\name{menu.mcrv.fittodata}
\alias{menu.mcrv.fittodata}
\title{Non-executable auxiliary function}
\usage{
menu.mcrv.fittodata(item,rriskModel,menuLevel=1)
}
\arguments{
\item{item}{...}

\item{rriskModel}{...}

\item{menuLevel}{...}
}
\keyword{items}

