% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelAuthorsClass-class}
\alias{modelAuthorsClass}
\alias{modelAuthorsClass-class}
\title{S4 class for representing 'rrisk' model authors}
\description{
S4 class for representing information of all the authors of a \code{rrisk} models
}
\examples{
\donttest{new("modelAuthorsClass")

new("modelAuthorsClass",authors=list(new("authorClass",name="Max Mustermann",
institution="Example institution",email="example@email.com")))}
}

