% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelBasicsClass-class}
\alias{modelBasicsClass}
\alias{modelBasicsClass-class}
\title{S4 class for representing 'rrisk' model basic descriptions}
\description{
S4 class for representing basic descriptions of a \code{rrisk} model
}
\examples{
\donttest{new("modelBasicsClass")}
}

