% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelClass-class}
\alias{modelClass}
\alias{modelClass-class}
\title{S4 class for representing 'rrisk' models}
\description{
S4 class for representing \code{rrisk} models
}
\examples{
 \donttest{new("modelClass")}
}

