% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelGlossaryClass-class}
\alias{modelGlossaryClass}
\alias{modelGlossaryClass-class}
\title{S4 class for representing 'rrisk' model glossary list}
\description{
S4 class for representing all the glossaries of a \code{rrisk} model
}
\examples{
\donttest{new("modelGlossaryClass")}
}

