% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelItemsClass-class}
\alias{modelItemsClass}
\alias{modelItemsClass-class}
\title{S4 class for representing all 'rrisk' model items}
\description{
S4 class for representing all the Items of \code{rrisk} models
}
\examples{
 \donttest{new("modelItemsClass")}
}

