% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelOutputClass-class}
\alias{modelOutputClass}
\alias{modelOutputClass-class}
\title{S4 class for representing 'rrisk' modelOutputClass}
\description{
S4 class for representing \code{rrisk} modelOutputClass
}
\section{Slots}{

\describe{
\item{\code{fullout.1d}}{list-object}

\item{\code{relaxout.1d}}{list-object}

\item{\code{fullout2d}}{character-vector}

\item{\code{uncitems.2d}}{character-vector}

\item{\code{OFname.2d}}{character}

\item{\code{summaries}}{character-vector}

\item{\code{runtime1d}}{numeric}

\item{\code{runtime2d}}{numeric}

\item{\code{OFcdfCI}}{numeric}
}}
\examples{
 \donttest{new("modelOutputClass")}
}

