% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelPartsClass-class}
\alias{modelPartsClass}
\alias{modelPartsClass-class}
\title{S4 class for representing 'rrisk' model parts}
\description{
S4 class for representing descriptions of all the parts of a \code{rrisk} model
}
\examples{
\donttest{new("modelPartsClass")}
}

