% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelReferencesClass-class}
\alias{modelReferencesClass}
\alias{modelReferencesClass-class}
\title{S4 class for representing 'rrisk' modeo references}
\description{
S4 class for representing all the references of a \code{rrisk} model
}
\examples{
 \donttest{new("modelReferencesClass")}
}

