% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelScoringClass-class}
\alias{modelScoringClass}
\alias{modelScoringClass-class}
\title{S4 class for representing 'rrisk' model scoring system}
\description{
S4 class for representing scoring systems of \code{rrisk} models
}
\examples{
 \donttest{new("modelScoringClass")}
}

