% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelUncertaintiesClass-class}
\alias{modelUncertaintiesClass}
\alias{modelUncertaintiesClass-class}
\title{S4 class for representing 'rrisk' model uncertainties}
\description{
S4 class for representing descriptions of all the uncertainties of a \code{rrisk} 
model
}
\examples{
\donttest{new(Class="modelUncertaintiesClass")}
}

