% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelValidationClass-class}
\alias{modelValidationClass}
\alias{modelValidationClass-class}
\title{S4 class for representing 'rrisk' model validities}
\description{
S4 class for representing all the validities of a \code{rrisk} model
}
\examples{
\donttest{new("modelValidationClass")

new("modelValidationClass",
validation=list(
new("validationClass",name="Model concept",
explanation=gsub(x="The model concept is a reasonable representation of 
the key processes that are to be considered to answer the risk question. 
The concept has been verified from a microbiological and modelling 
viewpoint.","\\n",replacement="")),
new("validationClass",
name="Model implementation and verification",
explanation=gsub(x="The correctness of the translation of the model
concept into the mathematical model has been checked. 
The appropriateness of the model parameterisation---and in 
particular of the choice of distributions for stochastic model 
input---has been commented in the item descriptions. The unit of 
observation of each model input quantity was carefully considered 
when combining inputs by mathematical expressions...","\\n",
replacement="")))) }
}

