% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{modelVersionClass-class}
\alias{modelVersionClass}
\alias{modelVersionClass-class}
\title{S4 class for representing 'rrisk' model version}
\description{
A S4 class for representing basic information of a \code{rrisk} model version
}
\examples{
\donttest{new("modelVersionClass")

new("modelVersionClass",status="Draft",minorupdate=0,majorupdate=1,
subtitle="Space for subtitle",editedby="Matthias Greiner")}
}

