\name{myvalue.stratumNames}
\alias{myvalue.stratumNames}
\title{Non-executable auxiliary function}
\usage{
  myvalue.stratumNames(item,rriskModel,menuLevel=1,stratumNames)
}
\arguments{
  \item{item}{...}

  \item{rriskModel}{...}

  \item{menuLevel}{...}

  \item{stratumNames}{...}
}
\description{
  Non-executable auxiliary function
}
\keyword{items}

