% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrisk-package.R
\docType{package}
\name{rrisk-package}
\alias{rrisk}
\alias{rrisk-package}
\title{rrisk: Quantitative risk modelling and auto-reporting in R.}
\description{
This R packages provides the whole functionality of the rrisk project
developed by Federal Institute for Risk Assessment. This tool supports the
risk assessor in definition, implementation and documantation of quantitative
risk models.
}
\keyword{package}

