% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rriskClass-class}
\alias{rriskClass}
\alias{rriskClass-class}
\title{S4 class for representing 'rrisk' session}
\description{
S4 class for representing \code{rrisk} session
}
\examples{
 \donttest{new("rriskClass")}
}

