% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,modelNameClass-method}
\alias{basicsClass-method}
\alias{graphClass-method}
\alias{itemClass-method}
\alias{modelGraphsClass-method}
\alias{modelItemsClass-method}
\alias{modelSettingsClass-method}
\alias{scoreClass-method}
\alias{show}
\alias{show,}
\alias{show,authorClass-method}
\alias{show,basicsClass-method}
\alias{show,glossaryClass-method}
\alias{show,graphClass-method}
\alias{show,itemClass-method}
\alias{show,modelAbbreviationsClass-method}
\alias{show,modelAuthorsClass-method}
\alias{show,modelBasicsClass-method}
\alias{show,modelClass-method}
\alias{show,modelCommentsClass-method}
\alias{show,modelConclusionsClass-method}
\alias{show,modelGlossaryClass-method}
\alias{show,modelGraphsClass-method}
\alias{show,modelItemsClass-method}
\alias{show,modelNameClass-method}
\alias{show,modelOutputClass-method}
\alias{show,modelPartsClass-method}
\alias{show,modelReferencesClass-method}
\alias{show,modelScoringClass-method}
\alias{show,modelSettingsClass-method}
\alias{show,modelUncertaintiesClass-method}
\alias{show,modelValidationClass-method}
\alias{show,modelVersionClass-method}
\alias{show,partClass-method}
\alias{show,scoreClass-method}
\alias{show,uncertClass-method}
\alias{show,validationClass-method}
\title{Show method for 'modelNameClass'}
\usage{
\S4method{show}{modelNameClass}(object)

\S4method{show}{modelVersionClass}(object)

\S4method{show}{authorClass}(object)

\S4method{show}{modelAuthorsClass}(object)

\S4method{show}{modelSettingsClass}(object)

\S4method{show}{basicsClass}(object)

\S4method{show}{modelBasicsClass}(object)

\S4method{show}{uncertClass}(object)

\S4method{show}{modelUncertaintiesClass}(object)

\S4method{show}{partClass}(object)

\S4method{show}{modelPartsClass}(object)

\S4method{show}{glossaryClass}(object)

\S4method{show}{modelGlossaryClass}(object)

\S4method{show}{modelAbbreviationsClass}(object)

\S4method{show}{validationClass}(object)

\S4method{show}{modelValidationClass}(object)

\S4method{show}{modelReferencesClass}(object)

\S4method{show}{modelCommentsClass}(object)

\S4method{show}{modelConclusionsClass}(object)

\S4method{show}{scoreClass}(object)

\S4method{show}{modelScoringClass}(object)

\S4method{show}{graphClass}(object)

\S4method{show}{modelGraphsClass}(object)

\S4method{show}{itemClass}(object)

\S4method{show}{modelItemsClass}(object)

\S4method{show}{modelOutputClass}(object)

\S4method{show}{modelClass}(object)
}
\arguments{
\item{object}{a \code{modelNameClass} object}

\item{object}{a \code{modelVersionClass} object}

\item{object}{a \code{authorClass} object}

\item{object}{a \code{modelAuthorsClass} object}

\item{object}{a \code{modelSettingsClass} object}

\item{object}{a \code{basicsClass} object}

\item{object}{a \code{modelBasicsClass} object}

\item{object}{a \code{uncertClass} object}

\item{object}{a \code{modelUncertaintiesClass} object}

\item{object}{a \code{partClass} object}

\item{object}{a \code{modelPartsClass} object}

\item{object}{a \code{glossaryClass} object}

\item{object}{a \code{modelGlossaryClass} object}

\item{object}{a \code{modelAbbreviationsClass} object}

\item{object}{a \code{validationClass} object}

\item{object}{a \code{modelValidationClass} object}

\item{object}{a \code{modelReferencesClass} object}

\item{object}{a \code{modelCommentsClass} object}

\item{object}{a \code{modelConclusionsClass} object}

\item{object}{a \code{scoreClass} object}

\item{object}{a \code{modelScoringClass} object}

\item{object}{a \code{graphClass} object}

\item{object}{a \code{modelGraphsClass} object}

\item{object}{a \code{itemClass} object}

\item{object}{a \code{modelItemsClass} object}

\item{object}{a \code{modelOutputClass} object}

\item{object}{a \code{modelClass} object}
}
\description{
Show method for \code{\linkS4class{modelNameClass}}

Show method for \code{\linkS4class{modelVersionClass}}

Show method for \code{\linkS4class{authorClass}}

Show method for \code{\linkS4class{modelAuthorsClass}}

Show method for \code{\linkS4class{modelSettingsClass}}

Show method for \code{\linkS4class{basicsClass}}

Show method for \code{\linkS4class{modelBasicsClass}}

Show method for \code{\linkS4class{uncertClass}}

Show method for \code{\linkS4class{modelUncertaintiesClass}}

Show method for \code{\linkS4class{partClass}}

Show method for \code{\linkS4class{modelPartsClass}}

Show method for \code{\linkS4class{glossaryClass}}

Show method for \code{\linkS4class{modelGlossaryClass}}

Show method for \code{\linkS4class{modelAbbreviationsClass}}

Show method for \code{\linkS4class{validationClass}}

Show method for \code{\linkS4class{modelValidationClass}}

Show method for \code{\linkS4class{modelReferencesClass}}

Show method for \code{\linkS4class{modelCommentsClass}}

Show method for \code{\linkS4class{modelConclusionsClass}}

Show method for \code{\linkS4class{scoreClass}}

Show method for \code{\linkS4class{modelScoringClass}}

Show method for \code{\linkS4class{graphClass}}

Show method for \code{\linkS4class{modelGraphsClass}}

Show method for \code{\linkS4class{itemClass}}

Show method for \code{\linkS4class{modelItemsClass}}

Show method for \code{\linkS4class{modelOutputClass}}

Show method for \code{\linkS4class{modelClass}}
}
\examples{
\donttest{show(new("modelNameClass"))}
\donttest{show(new("modelVersionClass"))}
\donttest{show(new("authorClass")) }
\donttest{show(new("modelAuthorsClass"))}
\donttest{show(new("modelSettingsClass"))}
\donttest{show(new("basicsClass"))}
\donttest{show(new("modelBasicsClass"))}
\donttest{show(new("uncertClass"))}
\donttest{show(new(Class="modelUncertaintiesClass"))}
\donttest{show(new("partClass"))}
\donttest{show(new("modelPartsClass"))}
\donttest{show(new("glossaryClass"))}
\donttest{show(new("modelGlossaryClass"))}
\donttest{show(new("modelAbbreviationsClass"))}
\donttest{show(new("validationClass"))}
\donttest{show(new("modelValidationClass"))}
 \donttest{show(new("modelReferencesClass"))}
 \donttest{show(new("modelCommentsClass"))}
 \donttest{show(new("modelConclusionsClass"))

new("modelConclusionsClass",
conclusions=list(gsub(x="This illustrative model predicts that the population
average probability of illness due to E.coli O157:H7 in child below 
age of three years attributable to consumption of one serving prepared 
from ground beef is around...","\\n",replacement=""),
gsub(x="The probabilistic modelling approach indicates that, taking into 
account variability and uncertainty, this probability could be much 
higher in rare cases...","\\n",replacement="")))}
 \donttest{show(new("scoreClass"))

new(Class="scoreClass",
notation="U1",
name="Plausibility",
values=c(0,1,2,3),
explanation=gsub(x="1=high (highly plausible despite absence of factual
evidence), 2=medium (plausible despite absence of factual evidence),
3=low (plausibility questionable or not assessed;
fictive or speculative assertion)","\\n",replacement=""))}
 \donttest{show(new("modelScoringClass"))}
 \donttest{show(new("graphClass"))}
 \donttest{show(new("modelGraphsClass"))}
 \donttest{show(new("itemClass"))}
 \donttest{show(new("modelItemsClass"))}
 \donttest{show(new("modelOutputClass"))    }
 \donttest{show(new("modelClass"))}
}

